remove old app scripts and templates
