--- app/scripts/views/app.js
@@ -1,15 +1,0 @@
-'use strict';
-
-var template = require('../../templates/app.hbs');
-var View = require('ampersand-view');
-
-var AppView = View.extend({
-	template: template,
-	render: function () {
-		this.renderWithTemplate(this.model);
-		document.body.appendChild(this.el);
-		return this;
-	}
-});
-
-module.exports = AppView;
