--- app/templates/alert.hbs
@@ -1,4 +1,0 @@
-<div class="alert alert-{{type}} {{#if dismissable}}alert-dismissible{{/if}}" role="alert">
-  {{#if dismissable}}<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>{{/if}}
-  <strong>{{heading}}</strong> {{message}}
-</div>
