--- app/templates/app.hbs
@@ -1,10 +1,0 @@
-<div class="main">
-	<div class="alert-container"></div>
-	<form action="" class="new-transaction">
-		<h2>Add a new transaction</h2>
-		<fieldset data-hook="field-container"></fieldset>
-		<button class="btn btn-primary" type="submit">Add</button>
-	</form>
-	<div class="stats"></div>
-	<div class="transactions"></div>
-</div>
