--- app/templates/stats.hbs
@@ -1,14 +1,0 @@
-<div class="stats">
-	<h2>Stats</h2>
-	<div class="averages">
-		<h3>Weekly Averages</h3>
-		<div class="stat">
-			{{#each averages}}
-				<div>
-					<label for="{{slug}}-avg">{{label}}</label>
-					<span id="{{slug}}-avg">{{money amount}}</span>
-				</div>
-			{{/each}}
-		</div>
-	</div>
-</div>
