--- app/templates/transactions.hbs
@@ -1,13 +1,0 @@
-<div class="transactions">
-	<div class="remove-transaction-modal modal fade">
-		<div class="modal-dialog modal-sm">
-			<div class="modal-content">
-				<div class="modal-header"><h4 class="modal-title">Are you sure you want to delete this transaction?</h4></div>
-				<div class="modal-footer">
-					<button type="button" class="btn btn-default" data-dismiss="modal">Go back</button>
-					<button type="button" class="btn btn-danger confirm-delete" data-id="">Delete</button>
-				</div>
-			</div>
-		</div>
-	</div>
-</div>
