--- app/templates/week.hbs
@@ -1,42 +1,0 @@
-<div class="weekly">
-	<h3>{{date 'MMM D' model.weekStart}} - {{date 'MMM D' model.weekEnd}}</h3>
-	<table class="table table-striped weekly-transactions">
-		<thead>
-			<tr>
-				<th>Day</th>
-				<th>Merchant</th>
-				<th>Amount</th>
-				<th>Source</th>
-				<th class="secondary">Desc.</th>
-				<th class="secondary">Category</th>
-				<th class="secondary"></th>
-			</tr>
-			<tr class="addition">
-				<th class="two-col">Desc.</th>
-				<th>Category</th>
-				<th></th>
-			</tr>
-		</thead>
-		<tbody>
-		</tbody>
-	</table>
-	<div class="summary">
-		<div class="week-total">
-			<h3>Week Total</h3>
-			<div class="stat">
-				<span class="value">{{money model.weekTotal}}</span>
-			</div>
-		</div>
-		<div class="by-category">
-			<h3>By category</h3>
-			<div class="stat">
-				{{#each stats.totals}}
-					<div class="cat-total">
-						<label for="{{slug}}-cat-total">{{label}}</label>
-						<span id="{{slug}}-cat-total">{{money amount}}</span>
-					</div>
-				{{/each}}
-			</div>
-		</div>
-	</div>
-</div>
